pyproject.toml
[tool.poetry]
name = "api_boilerplate"
version = "0.1.0"
description = ""
authors = ["Brad Anderson <brad.anderson58@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"
psycopg2 = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
